+++ reporting_system/urls.py
@@ -26,8 +26,8 @@ urlpatterns = [
          name='employee-logout'),
     path("employee/<int:user_id>",views.employeeview, name="employeeview"),
     path("report/",views.publicview, name="publicview"),
-    path("employee/<int:user_id>/flaw",user_view.add_flaw, name="add-flaw"),
-    path("employee/<int:user_id>/remove/<int:flaw>",user_view.remove_flaw, name="remove-flaw"),
-    path("employee/<int:user_id>/edit/<int:flaw>",user_view.edit_flaw, name="edit-flaw"),
+    path("employee/<int:user_id>/flaw",views.add_flaw, name="add-flaw"),
+    path("employee/<int:user_id>/remove/<int:flaw>",views.remove_flaw, name="remove-flaw"),
+    path("employee/<int:user_id>/edit/<int:flaw>",views.edit_flaw, name="edit-flaw"),
 
 ]
